fix hip path tracer

+++ src/hip_backend/kernels/constant_params.hip.h
@@ -2,7 +2,7 @@
 
 struct ConstantParams
 {
-    float depth;
+    uint32_t depth;
     uint32_t width;
     uint32_t height;
     uint32_t flip_y;
